Add normalization support for reward models
