+++ src/imitation/scripts/common/reward.py
@@ -1,12 +1,14 @@
 """Common configuration elements for reward network training."""
 
 import logging
-from typing import Any, Mapping, Optional, Type
+from typing import Any, Mapping, Type
 
 import sacred
 from stable_baselines3.common import vec_env
+from torch import nn
 
 from imitation.rewards import reward_nets
+from imitation.util import networks
 
 reward_ingredient = sacred.Ingredient("reward")
 logger = logging.getLogger(__name__)
@@ -20,12 +22,32 @@ def config():
     locals()  # quieten flake8
 
 
+@reward_ingredient.named_config
+def normalize_batchnorm():
+    net_kwargs = {"normalize_layer": nn.BatchNorm1d}  # noqa: F841
+
+
+@reward_ingredient.named_config
+def normalize_running():
+    net_kwargs = {"normalize_layer": networks.RunningNorm}  # noqa: F841
+
+
+@reward_ingredient.config_hook
+def config_hook(config, command_name, logger):
+    if config["reward"]["net_cls"] is None:
+        default_net = reward_nets.BasicRewardNet
+        if command_name == "airl":
+            default_net = reward_nets.BasicShapedRewardNet
+        return {"net_cls": default_net}
+    return {}
+
+
 @reward_ingredient.capture
 def make_reward_net(
     venv: vec_env.VecEnv,
-    net_cls: Optional[Type[reward_nets.RewardNet]],
-    net_kwargs: Optional[Mapping[str, Any]],
-) -> Optional[reward_nets.RewardNet]:
+    net_cls: Type[reward_nets.RewardNet],
+    net_kwargs: Mapping[str, Any],
+) -> reward_nets.RewardNet:
     """Builds a reward network.
 
     Args:
@@ -36,12 +58,10 @@ def make_reward_net(
     Returns:
         None if `reward_net_cls` is None; otherwise, an instance of `reward_net_cls`.
     """
-    if net_cls is not None:
-        net_kwargs = net_kwargs or {}
-        reward_net = net_cls(
-            venv.observation_space,
-            venv.action_space,
-            **net_kwargs,
-        )
-        logging.info(f"Reward network:\n {reward_net}")
-        return reward_net
+    reward_net = net_cls(
+        venv.observation_space,
+        venv.action_space,
+        **net_kwargs,
+    )
+    logging.info(f"Reward network:\n {reward_net}")
+    return reward_net
